requirements.txt
discord.py==2.3.2
Flask==3.0.0
PyNaCl==1.5.0
yt-dlp==2023.10.13